add useAudioSetup hook
